+++ docs/.vitepress/theme/components/blog/home/BlogHomePost.vue
@@ -1,8 +1,8 @@
 <script setup lang="ts">
 import { useData } from 'vitepress'
-import type { Post } from '../../composables/posts.data'
-import PostIcon from './PostIcon.vue'
-import PostAuthor from './PostAuthor.vue'
+import type { Post } from '../../../composables/posts.data'
+import PostIcon from '../PostIcon.vue'
+import BlogHomeAuthor from './BlogHomeAuthor.vue'
 
 const props = defineProps<{
   post: Post
@@ -30,24 +30,13 @@ const { site } = useData()
     </h2>
     <p class="mb-5 font-light" v-html="post.excerpt" />
     <div class="flex justify-between items-center">
-      <PostAuthor :name="post.author" />
+      <BlogHomeAuthor :name="post.author" />
       <a
         :href="`${site.base}blog${post.href}`"
         class="inline-flex items-center font-medium hover:text-[color:var(--vp-c-brand-dark)]"
       >
         Read more
-        <svg
-          class="ml-2 w-4 h-4"
-          fill="currentColor"
-          viewBox="0 0 20 20"
-          xmlns="http://www.w3.org/2000/svg"
-        >
-          <path
-            fill-rule="evenodd"
-            d="M10.293 3.293a1 1 0 011.414 0l6 6a1 1 0 010 1.414l-6 6a1 1 0 01-1.414-1.414L14.586 11H3a1 1 0 110-2h11.586l-4.293-4.293a1 1 0 010-1.414z"
-            clip-rule="evenodd"
-          />
-        </svg>
+        <div class="i-[bx/right-arrow-alt] ml-2" />
       </a>
     </div>
   </article>
