+++ docs/.vitepress/theme/layout/PostBottom.vue
@@ -1,6 +1,7 @@
 <script setup lang="ts">
-import PostLinks from '../PostLinks.vue'
+import PostLinks from '../components/blog/post/PostLinks.vue'
 </script>
+
 <template>
   <PostLinks inside-doc />
 </template>
