+++ docs/blog/index.md
@@ -2,6 +2,4 @@
 layout: home
 ---
 
-<Posts/>
-
-
+<BlogHomeIndex/>
